feat: split js into modules
