+++ index.html
@@ -110,6 +110,6 @@
     </main>
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.1/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
-    <script src="./script.js" type="module"></script>
+    <script src="./js/script.js" type="module"></script>
 </body>
 </html>
